+++ Airship/AirshipCore/Tests/UAirshipTest.m
@@ -66,7 +66,12 @@
     id component1 = [self mockForProtocol:@protocol(UAComponent)];
     id component2 = [self mockForProtocol:@protocol(UAComponent)];
     id component3 = [self mockForProtocol:@protocol(UAComponent)];
-    
+
+    id mockDelegate = [self mockForProtocol:@protocol(UADeepLinkDelegate)];
+    [[mockDelegate reject] receivedDeepLink:deepLink completionHandler:OCMOCK_ANY];
+
+    [UAirship shared].deepLinkDelegate = mockDelegate;
+
     [[[component1 expect] andReturnValue:@(NO)] deepLink:deepLink];
     [[[component2 expect] andReturnValue:@(YES)] deepLink:deepLink];
     [[component3 reject] deepLink:deepLink];
@@ -85,6 +90,35 @@
     [component1 verify];
     [component2 verify];
     [component3 verify];
+    [mockDelegate verify];
+}
+
+- (void)testUAirshipComponentsDeepLinksFallbackDelegate {
+    NSURL *deepLink = [NSURL URLWithString:@"uairship://some-deep-link"];
+    id component1 = [self mockForProtocol:@protocol(UAComponent)];
+    [[[component1 expect] andReturnValue:@(NO)] deepLink:deepLink];
+
+    id mockDelegate = [self mockForProtocol:@protocol(UADeepLinkDelegate)];
+    [[[mockDelegate expect] andDo:^(NSInvocation *invocation) {
+        void (^completionHandler)(void);
+        [invocation getArgument:&completionHandler atIndex:3];
+        completionHandler();
+    }] receivedDeepLink:deepLink completionHandler:OCMOCK_ANY];
+
+    self.airshipInstance.components = @[component1];
+    [UAirship shared].deepLinkDelegate = mockDelegate;
+
+    XCTestExpectation *deepLinked = [self expectationWithDescription:@"Deep linked"];
+    [[UAirship shared] deepLink:deepLink completionHandler:^(BOOL result) {
+        XCTAssertTrue(result);
+        [deepLinked fulfill];
+    }];
+
+    // Wait for the test expectations
+    [self waitForTestExpectations];
+
+    [component1 verify];
+    [mockDelegate verify];
 }
 
 - (void)testUAirshipComponentsDeepLinksAlwaysReturnsTrue {